python: move AltField class to flux.util
Problem: The AltField class is only present in flux-resource.py,
but this class could be useful in other commands.

Move it to flux.util.
